+++ scripts/tasks/linc.js
@@ -8,22 +8,9 @@
 'use strict';
 
 const lintOnFiles = require('../eslint');
-const execFileSync = require('child_process').execFileSync;
-const mergeBase = execFileSync('git', ['merge-base', 'HEAD', 'master'], {
-  stdio: 'pipe',
-  encoding: 'utf-8',
-}).trim();
-const changedFiles = execFileSync(
-  'git',
-  ['diff', '--name-only', '--diff-filter=ACMRTUB', mergeBase],
-  {
-    stdio: 'pipe',
-    encoding: 'utf-8',
-  }
-)
-  .trim()
-  .toString()
-  .split('\n');
+const listChangedFiles = require('../shared/listChangedFiles');
+
+const changedFiles = [...listChangedFiles()];
 const jsFiles = changedFiles.filter(file => file.match(/.js$/g));
 
 const report = lintOnFiles(jsFiles);
